+++ R/rows.R
@@ -1,0 +1,10 @@
+
+#' @export
+rows <- function(.data) {
+  UseMethod("rows")
+}
+
+#' @export
+rows.data.frame <- function(.data) {
+  rows_impl(.data)
+}