+++ src/RcppExports.cpp
@@ -205,6 +205,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// rows_impl
+List rows_impl(DataFrame data);
+RcppExport SEXP _dplyr_rows_impl(SEXP dataSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< DataFrame >::type data(dataSEXP);
+    rcpp_result_gen = Rcpp::wrap(rows_impl(data));
+    return rcpp_result_gen;
+END_RCPP
+}
 // as_regular_df
 DataFrame as_regular_df(DataFrame df);
 RcppExport SEXP _dplyr_as_regular_df(SEXP dfSEXP) {
@@ -719,6 +730,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_dplyr_n_distinct_multi", (DL_FUNC) &_dplyr_n_distinct_multi, 2},
     {"_dplyr_filter_impl", (DL_FUNC) &_dplyr_filter_impl, 2},
     {"_dplyr_slice_impl", (DL_FUNC) &_dplyr_slice_impl, 2},
+    {"_dplyr_rows_impl", (DL_FUNC) &_dplyr_rows_impl, 1},
     {"_dplyr_as_regular_df", (DL_FUNC) &_dplyr_as_regular_df, 1},
     {"_dplyr_ungroup_grouped_df", (DL_FUNC) &_dplyr_ungroup_grouped_df, 1},
     {"_dplyr_test_grouped_df", (DL_FUNC) &_dplyr_test_grouped_df, 1},